Set testing program done

+++ src/set/vclock.rs
@@ -8,13 +8,13 @@ pub enum VClockCompareResult {
   CONCURRENT,
 }
 
-#[derive(Clone)]
+#[derive(Clone, Debug)]
 pub struct VClock {
   pub clock: HashMap<String, i32>,
 }
 
 impl VClock {
-  pub fn new(total_replicas: i32) -> Self {
+  pub fn new(total_replicas: usize) -> Self {
       let mut tmp = Self {
           clock: HashMap::new(),
       };
@@ -74,7 +74,7 @@ impl VClock {
       // let temp = VClock::new_with_clock(self.clock.clone());
       let temp_map = clk.clock.clone();
       let mut temp = VClock::new_with_clock(temp_map);
-      // temp.increment(id);
+      
       temp.clock
           .insert(id.clone(), temp.clock.get(id).unwrap() + 1);
       temp
